+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+### Add
+
+- Use a fleet adapter custom package
+- Mqtt variables can be configured
+- Navigation is considered completed when the distance is less than 0.5 or when the topic result indicates it.
+- If the robot position is not receive when the fleet adapter start the system is still running.
+
 ## [0.1.0] - 2023-05-04
 
 ### Removed